chore(build): make build fail if test fails

+++ scripts/test.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/sh -e
 
 # Get Test dependencies
 go get github.com/axw/gocov/gocov
@@ -12,9 +12,14 @@ go get github.com/modocache/gover
 echo "mode: count" > profile.cov
 
 # Standard go tooling behavior is to ignore dirs with leading underscors
-for dir in $(find . -maxdepth 10 -not -path './.git*' -not -path './vendor*' -not -path '*/_*' -type d); do
+for dir in $(find . -maxdepth 10 -not -path './.git*' -not -path './vendor*' -not -path './examples/*' -not -path './pkg' -not -path './bin' -not -path './scripts' -not -path '*/_*' -type d); do
   if ls $dir/*.go &> /dev/null; then
     go test -covermode=count -coverprofile=$dir/profile.tmp $dir
+    if [ $? = 1 ]; then
+      echo "Test failure, exiting"
+      exit 1
+    fi
+        
     if [ -f $dir/profile.tmp ]; then
     	cat $dir/profile.tmp | tail -n +2 >> profile.cov
     	rm $dir/profile.tmp
